# doc/api-doc.yml
openapi: 3.0.0
info:
  title: 武大论坛接口文档
  version: 1.0.0
servers:
  - url: whuerbbs.cn/v1
paths:
...
  /comments:
    get:
      summary: 分页获取评论
      parameters:
        - in: query
          name: post_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentListVO'
    post:
      summary: 提交评论
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                post_id:
                  type: integer
                content:
                  type: string
                  maxLength: 300
                parent_id:
                  type: integer
              required:
                - post_id
                - content
      responses:
        200:
          $ref: '#/components/responses/Ok'
  /sub_comments:
    get:
      summary: 分页获取某评论的子评论
      parameters:
        - in: query
          name: comment_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentListVO'
  /hot_comments:
    get:
      summary: 获取热门评论
      parameters:
        - in: query
          name: post_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentListVO'
  /comments/{comment_id}:
    get:
      summary: 获取评论详情
      parameters:
        - in: path
          name: comment_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentListVO'
    delete:
      summary: 删除评论
      parameters:
        - in: path
          name: comment_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Ok'
...
  /comments/{comment_id}/like:
    get:
      summary: 点赞评论
      parameters:
        - in: path
          name: comment_id
          required: true
          schema:
            type: integer
      responses:
        200:
          $ref: '#/components/responses/Ok'
  /comments/{comment_id}/cancel_like:
    get:
      summary: 取消评论点赞
      parameters:
        - in: path
          name: comment_id
          required: true
          schema:
            type: integer
      responses:
        200:
          $ref: '#/components/responses/Ok'
...
  /posts:
    get:
      summary: 获取全部帖子（全部、问点事、一起玩共用）
      parameters:
        - in: query
          name: board
          required: true
          description: 版块
          schema:
            type: integer
            default: null
        - in: query
          name: topic_id
          description: 话题id
          schema:
            type: integer
            default: null
        - in: query
          name: hot
          description: 是否热门
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListVO'
    post:
      summary: 发布帖子
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/PostDTO'
                - type: object
                  properties:
                    topic_id:
                      type: integer
                    board:
                      type: integer
      responses:
        200:
          $ref: '#/components/responses/Ok'
  /posts/{post_id}:
    get:
      summary: 获取帖子详情
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostVO'
  /posts/{post_id}/like:
    get:
      summary: 点赞帖子
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
      responses:
        200:
          $ref: '#/components/responses/Ok'
  /posts/{post_id}/cancel_like:
    get:
      summary: 取消帖子点赞
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
      responses:
        200:
          $ref: '#/components/responses/Ok'
  /posts/{post_id}/collect:
    get:
      summary: 收藏帖子
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
      responses:
        200:
          $ref: '#/components/responses/Ok'
  /posts/{post_id}/cancel_collect:
    get:
      summary: 取消收藏帖子
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
      responses:
        200:
          $ref: '#/components/responses/Ok'
...
  /anonymous_posts:
    get:
      summary: 获取全部匿名帖子
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListVO'
    post:
      summary: 发表匿名帖子
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/PostDTO'
                - type: object
                  properties:
                    topic_id:
                      type: integer
                    anonymous_name:
                      type: string
      responses:
        200:
          $ref: '#/components/responses/Ok'
  /anonymous_posts/{anonymous_post_id}:
    get:
      summary: 获取匿名帖子详情
      parameters:
        - in: path
          name: anonymous_post_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PostVO'
                  - type: object
                    properties:
                      anonymous_name:
                        type: string
...
components:
  responses:
    Ok:
      description: 成功
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
              message:
                type: integer
...
  schemas:
    AttitudeStatus:
      type: integer
      enum: [0, 1]
      description: >
        Sort order:
         * `0` - 默认
         * `1` - 点赞
...
    Board:
      type: integer
      enum: [1, 2, 3, 4, 5]
      description: >
        Sort order:
          * `1` - 新鲜事
          * `2` - 二手区
          * `3` - 问点事
          * `4` - 匿名墙
          * `5` - 一起玩
...
    UserVO:
      type: object
      properties:
        id:
          type: string
        nickname:
          type: string
        avatar_url:
          type: string
...
    TopicVO:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
    PostDTO:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        images:
          type: array
          items:
            type: string
    PostListVO:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        board:
          $ref: '#/components/schemas/Board'
        image:
          type: string
        like_count:
          type: integer
        comment_count:
          type: integer
        created_at:
          type: string
          format: 'date-time'
        last_active_at:
          type: string
          format: 'date-time'
        created_by:
          $ref: '#/components/schemas/UserVO'
        topics:
          type: array
          items:
            $ref: '#/components/schemas/TopicVO'
    PostVO:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        board:
          $ref: '#/components/schemas/Board'
        images:
          type: array
          items:
            type: string
        like_count:
          type: integer
        comment_count:
          type: integer
        last_active_at:
          type: integer
        attitude_status:
          $ref: '#/components/schemas/AttitudeStatus'
        created_by:
          $ref: '#/components/schemas/UserVO'
        topics:
          type: array
          items:
            $ref: '#/components/schemas/TopicVO'
        hot:
          type: boolean
        collected:
          type: boolean
        can_delete:
          type: boolean
    CommentListVO:
      allOf:
        - $ref: '#/components/schemas/CommentVO'
        - properties:
            sub_comments:
              type: array
              items:
                $ref: '#/components/schemas/CommentVO'
            sub_comment_count:
              type: integer
    CommentVO:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        like_count:
          type: integer
        created_at:
          type: string
          format: 'date-time'
        created_by:
          $ref: '#/components/schemas/UserVO'
        attitude_status:
          $ref: '#/components/schemas/AttitudeStatus'
        can:delete:
          type: boolean
...